fix: fix sandbox timer module bug
